Implement context api useContext state management updates for react tutorial router project
