+++ data/db.json
@@ -15,8 +15,8 @@
     {
       "id": "4",
       "title": "My Fourth Post",
-      "datetime": "February 07, 2024 9:29:11 AM",
-      "body": "Some more testing paragraphs."
+      "datetime": "February 11, 2024 8:02:42 AM",
+      "body": "Some more testing words"
     }
   ]
 }
