+++ src/EditPost.js
@@ -1,10 +1,15 @@
 import React from 'react';
-import { useEffect } from 'react';
-import { useParams, Link } from 'react-router-dom';
+import { useState, useEffect, useContext } from 'react';
+import { useParams, Link, useHistory } from 'react-router-dom';
+import api from './api/posts';
+import { format } from 'date-fns';
+import DataContext from './context/DataContext';
 
-const EditPost = ({
-    posts, handleEdit, editBody, setEditBody, editTitle, setEditTitle
-}) => {
+const EditPost = () => {
+    const [editTitle, setEditTitle] = useState('');
+    const [editBody, setEditBody] = useState('');
+    const { posts, setPosts } = useContext(DataContext);
+    const history = useHistory();
     const { id } = useParams();
     const post = posts.find(post => (post.id).toString() === id);
     
@@ -14,6 +19,20 @@ const EditPost = ({
             setEditBody(post.body);
         }
     }, [post, setEditTitle, setEditBody]);
+
+    const handleEdit = async (id) => {
+        const datetime = format(new Date(), 'MMMM dd, yyyy pp');
+        const updatedPost = {id, title: editTitle, datetime, body: editBody };
+        try {
+            const response = await api.put(`/posts/${id}`, updatedPost);
+            setPosts(posts.map(post => post.id === id ? { ...response.data } : post));
+            setEditTitle('');
+            setEditBody('');
+            history.push('/');
+        } catch (err) {
+            console.log(`Error: ${err.message}`);
+        }
+    }
 
     return (
         <main className="NewPost">
@@ -34,7 +53,7 @@ const EditPost = ({
                             id="postBody" 
                             required 
                             value={editBody} 
-                            onChange={(e) => setEditBody(e.target.value)}
+                            onChange={(e) => setEditBody(e.target.value)} 
                         />
                         <button type="submit" onClick={() => handleEdit(post.id)}>Submit</button>
                     </form>
