+++ src/Header.js
@@ -1,7 +1,10 @@
 import React from 'react';
 import { FaLaptop, FaTabletAlt, FaMobileAlt } from 'react-icons/fa';
+import useWindowSize from './hooks/useWindowSize';
 
-const Header = ({ title, width }) => {
+const Header = ({ title }) => {
+    const { width } = useWindowSize();
+    
     return (
         <header className="Header">
             <h1>{title}</h1>
