+++ src/Nav.js
@@ -1,7 +1,11 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import { useContext } from 'react';
+import DataContext from './context/DataContext';
 
-const Nav = ({ search, setSearch }) => {
+const Nav = () => {
+    const { search, setSearch } = useContext(DataContext);
+
     return (
         <nav className="Nav">
             <form className="searchForm" onSubmit={(e) => e.preventDefault()}>
