+++ src/PostPage.js
@@ -1,9 +1,25 @@
 import React from 'react';
-import { useParams, Link } from 'react-router-dom';
+import { useParams, Link, useHistory } from 'react-router-dom';
+import { useContext } from 'react';
+import api from './api/posts';
+import DataContext from './context/DataContext';
 
-const PostPage = ({ posts, handleDelete }) => {
+const PostPage = () => {
+    const { posts, setPosts } = useContext(DataContext);
     const { id } = useParams();
+    const history = useHistory();
     const post = posts.find(post => (post.id).toString() === id);
+
+    const handleDelete = async (id) => {
+        try {
+            await api.delete(`/posts/${id}`);
+            const postsList = posts.filter(post => post.id !== id);
+            setPosts(postsList);
+            history.push('/');
+        } catch (err) {
+            console.log(`Error: ${err.message}`);
+        }
+    }
 
     return (
         <main className="PostPage">
